+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 ### Added
+- MTL material editor with preview
+- Material editor settings:
+  - Default material preview mesh (cube, cylinder, sphere)
+  - Default material type (Phong or PBR)
 - New OBJ file action
 - New MTL file action
 - Use JVM toolchain for configuring source/target compilation compatibility